delete unused import and use logger.logf in example

+++ src/org/sosy_lab/java_smt/example/IndependentInterpolation.java
@@ -112,7 +112,7 @@ public final class IndependentInterpolation {
     Preconditions.checkState(unsat, "The example for interpolation should be UNSAT");
 
     BooleanFormula itp = prover.getInterpolant(ImmutableList.of(ip1));
-    logger.log(
+    logger.logf(
         Level.INFO,
         String.format(
             "Interpolation Result:%n"
@@ -155,7 +155,7 @@ public final class IndependentInterpolation {
 
     BooleanFormula itp = prover.getInterpolant(ImmutableList.of(ip1));
     logger.log(Level.INFO, "Interpolants are:", itp);
-    logger.log(
+    logger.logf(
         Level.INFO,
         String.format(
             "Interpolation Result:%n"
